refactor(PaintControls): factor out FillBetweenControls
